Add [test groups] pages/Task/Tests
